% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rORM-package.R
\docType{package}
\name{rORM-package}
\alias{rORM}
\alias{rORM-package}
\title{rORM: Database Object Relational Mapper for R Language using DBI}
\description{
The rORM package give you the possibility to generate R code which provides R6 objects for each DB table. This objects works as models which can be used to perform basic CRUD operations on the DB.
}
\keyword{internal}
