% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{rorm_cleanup_database_from_testdata}
\alias{rorm_cleanup_database_from_testdata}
\title{DB helper function for package examples and package testing which clean up
the test DB setup made with `rorm_initialize_database_with_testdata`.}
\usage{
rorm_cleanup_database_from_testdata(con)
}
\arguments{
\item{con}{DB connection}
}
\description{
DB helper function for package examples and package testing which clean up
the test DB setup made with `rorm_initialize_database_with_testdata`.
}
\examples{
rorm_cleanup_database_from_testdata(con)
}
