% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{rorm_connect_to_test_db}
\alias{rorm_connect_to_test_db}
\title{DB helper function for package examples and package testing which creates
a PostgreSQL DB connection to a default database. Credentials are hard coded
because it is only for demo purposes. Everyone can create their own DB
connection straight forward with the `DBI` package.}
\usage{
rorm_connect_to_test_db(type = c("sqlite", "postgresql")[1])
}
\arguments{
\item{type}{which type of driver should be used; one of
("sqlite", "postgresql").}
}
\value{
DBI DB connection
}
\description{
DB helper function for package examples and package testing which creates
a PostgreSQL DB connection to a default database. Credentials are hard coded
because it is only for demo purposes. Everyone can create their own DB
connection straight forward with the `DBI` package.
}
\examples{
rorm_connect_to_test_db(type = "sqlite")
}
