% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{rorm_extract_postgresql_structure_table}
\alias{rorm_extract_postgresql_structure_table}
\title{Given an existing DB Table which existing using the given PostgreSQL DB
connection (`con`) this function returns all table fields and all keys
if any exist.}
\usage{
rorm_extract_postgresql_structure_table(con, table_name)
}
\arguments{
\item{con}{DB connection}

\item{table_name}{Name of DB table}
}
\value{
list containing the `primary_keys` and `fields` of the table.
}
\description{
Given an existing DB Table which existing using the given PostgreSQL DB
connection (`con`) this function returns all table fields and all keys
if any exist.
}
\examples{
rorm_extract_postgresql_structure_table(con, "account")

}
