% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generator.R
\name{rorm_generate_code}
\alias{rorm_generate_code}
\title{Generate the code to map (PostgreSQL) DB tables to a RORM model.}
\usage{
rorm_generate_code(
  con,
  prefix = "RORM",
  verbose = FALSE,
  db_connection_name = "con"
)
}
\arguments{
\item{con}{DB connection}

\item{prefix}{The prefix of the class and model variables, default is RORM}

\item{verbose}{Whether the model runs in verbose mode and log all
sql statements}

\item{db_connection_name}{The name of the DBI connection variable used to
connect to the DB}
}
\value{
R code in string format which can be evaluated
}
\description{
Generate the code to map (PostgreSQL) DB tables to a RORM model.
}
\examples{
con <- rorm_connect_to_test_db(type = "sqlite")
rorm_generate_code(con)

rorm_generate_code(con, prefix = "DB")
}
