% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_generator.R
\name{rorm_generate_code_to_file}
\alias{rorm_generate_code_to_file}
\title{Generate the RORM mapper classes / models codes and write it to a file
which can be included in the R project}
\usage{
rorm_generate_code_to_file(
  con,
  prefix = "RORM",
  verbose = FALSE,
  db_connection_name = "con",
  filepath = "rorm_models.R"
)
}
\arguments{
\item{con}{DB connection}

\item{prefix}{The prefix of the class and model variables, default is RORM}

\item{verbose}{Whether the model runs in verbose mode and log all
sql statements}

\item{db_connection_name}{The name of the DBI connection variable used to
connect to the DB}

\item{filepath}{Path to source code}
}
\value{
writes the R code to a file
}
\description{
Generate the RORM mapper classes / models codes and write it to a file
which can be included in the R project
}
\examples{
rorm_generate_code_to_file(con)
}
