% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{rorm_initialize_database_with_testdata}
\alias{rorm_initialize_database_with_testdata}
\title{DB helper function for package examples and package testing which uses the
example R datasets to create a PostgreSQL Database for testing}
\usage{
rorm_initialize_database_with_testdata(con)
}
\arguments{
\item{con}{DB connection}
}
\description{
DB helper function for package examples and package testing which uses the
example R datasets to create a PostgreSQL Database for testing
}
\examples{
rorm_initialize_database_with_testdata(con)
}
