% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorm.R
\docType{data}
\name{rorm_logger}
\alias{rorm_logger}
\title{The package internal logger}
\format{
An object of class \code{Logger} of length 1.
}
\usage{
rorm_logger
}
\description{
The package internal logger
}
\keyword{datasets}
