% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorm.R
\name{RORMPostgreSQLBaseClass}
\alias{RORMPostgreSQLBaseClass}
\title{Base class to connect to a specific given (PostgreSQL) DB table to run
standardized CRUD operations, like create, read, update and delete on this
table.
This base class is mainly the core to the RORM package and will be inherited
using the code generator to configure the ORM mapper for each table.}
\description{
Base class to connect to a specific given (PostgreSQL) DB table to run
standardized CRUD operations, like create, read, update and delete on this
table.
This base class is mainly the core to the RORM package and will be inherited
using the code generator to configure the ORM mapper for each table.

Base class to connect to a specific given (PostgreSQL) DB table to run
standardized CRUD operations, like create, read, update and delete on this
table.
This base class is mainly the core to the RORM package and will be inherited
using the code generator to configure the ORM mapper for each table.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fields}}{fields of the connected table.}

\item{\code{key}}{if exists, the keys of the table, could be more if a union is in place.}

\item{\code{table_name}}{The name of the table.}

\item{\code{key_setting}}{The type of keys used (PRIMARY, UNION, ...)}

\item{\code{db_connection}}{DBI DB connection.
Initialize the Class (new method) and connect the ORM model to the DB}

\item{\code{verbose}}{for logging switch verbose to TRUE.}

\item{\code{dry_run}}{for not doing any DB operations switch dry_run to TRUE.
Delete a DB entry by key}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RORMPostgreSQLBaseClass-new}{\code{RORMPostgreSQLBaseClass$new()}}
\item \href{#method-RORMPostgreSQLBaseClass-delete}{\code{RORMPostgreSQLBaseClass$delete()}}
\item \href{#method-RORMPostgreSQLBaseClass-get}{\code{RORMPostgreSQLBaseClass$get()}}
\item \href{#method-RORMPostgreSQLBaseClass-all}{\code{RORMPostgreSQLBaseClass$all()}}
\item \href{#method-RORMPostgreSQLBaseClass-insert}{\code{RORMPostgreSQLBaseClass$insert()}}
\item \href{#method-RORMPostgreSQLBaseClass-update}{\code{RORMPostgreSQLBaseClass$update()}}
\item \href{#method-RORMPostgreSQLBaseClass-clone}{\code{RORMPostgreSQLBaseClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RORMPostgreSQLBaseClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-RORMPostgreSQLBaseClass-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RORMPostgreSQLBaseClass$new(con)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con}}{The DB connection}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ORM model object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RORMPostgreSQLBaseClass-delete"></a>}}
\if{latex}{\out{\hypertarget{method-RORMPostgreSQLBaseClass-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RORMPostgreSQLBaseClass$delete(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The key value what to delete.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Number of entries deleted
Method to load a row by key.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RORMPostgreSQLBaseClass-get"></a>}}
\if{latex}{\out{\hypertarget{method-RORMPostgreSQLBaseClass-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RORMPostgreSQLBaseClass$get(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The key value for what to filter for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The row which match the key.
Method to load the entire table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RORMPostgreSQLBaseClass-all"></a>}}
\if{latex}{\out{\hypertarget{method-RORMPostgreSQLBaseClass-all}{}}}
\subsection{Method \code{all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RORMPostgreSQLBaseClass$all()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Entire table
Method to insert new table content.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RORMPostgreSQLBaseClass-insert"></a>}}
\if{latex}{\out{\hypertarget{method-RORMPostgreSQLBaseClass-insert}{}}}
\subsection{Method \code{insert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RORMPostgreSQLBaseClass$insert(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{contains the new content of the DB table. Potential values for the key columns will be deleted if the table has a primary key and overwritten by the primary key logic of the DB.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Number of entries inserted.
Method to update existing table content.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RORMPostgreSQLBaseClass-update"></a>}}
\if{latex}{\out{\hypertarget{method-RORMPostgreSQLBaseClass-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RORMPostgreSQLBaseClass$update(key, df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The column value which content should be updated. If the table
has a primary key, then only a value can be provided, otherwise
a named vector is required. Does the table has no keys at all,
then any named vector will be used as WHERE argument.
This can lead to multiple row updates.}

\item{\code{df}}{The new data as data.frame.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Number of entries updated.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RORMPostgreSQLBaseClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RORMPostgreSQLBaseClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RORMPostgreSQLBaseClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
