% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorm.R
\docType{data}
\name{RORMPostgreSQLKeySetting}
\alias{RORMPostgreSQLKeySetting}
\title{RORM constant to define different table key settings like UNION or PRIMARY}
\format{
An object of class \code{list} of length 3.
}
\usage{
RORMPostgreSQLKeySetting
}
\description{
RORM constant to define different table key settings like UNION or PRIMARY
}
\keyword{datasets}
