% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{str_camel_case}
\alias{str_camel_case}
\title{General string helper function convert mainly snake_case strings which are
mainly used in DB tables to a CamelCase format.}
\usage{
str_camel_case(string)
}
\arguments{
\item{string}{string, i.e. better snake_case string}
}
\value{
camel case string.
}
\description{
kebab-case strings are also handles correctly, all other formats will not fails
but could lead to unexpected format.
}
\examples{
str_camel_case("table_name") # snake_case
str_camel_case("yet-another-table-name") # kebab-case
str_camel_case("CamelCaseTable") # already in CamelCase will be destroyed
# Spaces will be handled as snake case, but the assumption is that a DB table
# has no spaces.
str_camel_case("table name with spaces")
}
