% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{vector_to_R_code}
\alias{vector_to_R_code}
\title{General string helper converting a R vector to correct R code which can be
executed.}
\usage{
vector_to_R_code(vector)
}
\arguments{
\item{vector}{arbitrary R vector}
}
\value{
string which can be parsed to R code (see examples)
}
\description{
General string helper converting a R vector to correct R code which can be
executed.
}
\examples{
vector_to_R_code(1:10)
vector_to_R_code("Hello World")
vector_to_R_code(c(1, 2, 4, 5, "6"))
eval(parse(text = paste0(vector_to_R_code(1:10), " + 11")))
}
